Move intellisense provider to standalone file
